change c++ headers extension to hpp
