+++ Examples/FitSine/Data.cpp
@@ -17,7 +17,7 @@
 * along with DNest3. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "Data.h"
+#include "Data.hpp"
 #include <fstream>
 #include <iostream>
 #include <algorithm>
