+++ Examples/FitSine/Data.hpp
