+++ src/RandomNumberGenerator.cpp
@@ -17,7 +17,7 @@
 * along with DNest3. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "RandomNumberGenerator.h"
+#include "RandomNumberGenerator.hpp"
 #include <gsl/gsl_randist.h>
 #include <cmath>
 #include <cassert>
